% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.birth.counts.distn.R
\docType{data}
\name{sim.birth.counts.distn}
\alias{sim.birth.counts.distn}
\title{Simulated data on age-specific birth counts, with distribution perturbed.}
\format{An array with dimensions \code{"age"}, \code{"region"},
\code{"time"}, and \code{"iteration"}.}
\usage{
sim.birth.counts.distn
}
\description{
Simulated birth counts for 5-year age groups, 30 regions, and 3 time periods.
}
\details{
The rates were obtained by replacing counts from
\code{\link{sim.birth.counts.base}} in regions 25-30 in period 3
with Poisson draws with expected values equal to
\code{\link{sim.birth.rates.distn}} multiplied by
\code{\link{sim.fem.popn}}.
}
\seealso{
The data are designed for use with \code{\link{sim.fem.popn}},
\code{\link{sim.birth.rates.base}}, \code{\link{sim.birth.rates.distn}},
\code{\link{sim.birth.rates.level}}, \code{\link{sim.birth.counts.base}},
\code{\link{sim.birth.counts.level}}, and \code{\link{sim.birth.standard}}.
}
\keyword{datasets}
