% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.birth.rates.base.R
\docType{data}
\name{sim.birth.rates.base}
\alias{sim.birth.rates.base}
\title{Simulated data on age-specific birth rates.}
\format{An array with dimensions \code{"age"}, \code{"region"},
\code{"time"}, and \code{"iteration"}.}
\usage{
sim.birth.rates.base
}
\description{
Simulated birth rates for 5-year age groups, 30 regions, and 3 time periods.
}
\details{
The rates were obtained by taking age-specific fertility rates for
New Zealand in 2006, and multiplying by random region and time effects.
}
\seealso{
The data are designed for use with \code{\link{sim.fem.popn}},
\code{\link{sim.birth.rates.distn}}, \code{\link{sim.birth.rates.level}},
\code{\link{sim.birth.counts.base}}, \code{\link{sim.birth.counts.distn}},
\code{\link{sim.birth.counts.level}}, and \code{\link{sim.birth.standard}}.
}
\keyword{datasets}

