% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.birth.standard.R
\docType{data}
\name{sim.birth.standard}
\alias{sim.birth.standard}
\title{Logical array showing whether simulated data on births drawn from base
distribution.}
\format{An array with dimensions \code{"age"}, \code{"region"},
and \code{"time"}.}
\usage{
sim.birth.standard
}
\description{
Logical array equal to \code{FALSE} in regions 25-30 in period 3,
and \code{TRUE} otherwise.
}
\seealso{
The data are designed for use with \code{\link{sim.fem.popn}},
\code{\link{sim.birth.rates.base}}, \code{\link{sim.birth.rates.distn}},
\code{\link{sim.birth.rates.level}}, \code{\link{sim.birth.counts.base}},
\code{\link{sim.birth.counts.distn}}, and \code{\link{sim.birth.counts.level}}.
}
\keyword{datasets}

