% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fem.popn.R
\docType{data}
\name{sim.fem.popn}
\alias{sim.fem.popn}
\title{Simulated data on women of childbearing age.}
\format{An array with dimensions \code{"age"}, \code{"region"},
and \code{"time"}.}
\usage{
sim.fem.popn
}
\description{
Simulated numbers of women of childbearin age, by 5-year age group,
30 regions, and 3 time period.  Every cell in the the array has a
value of 300.
}
\seealso{
The data are designed for use with \code{\link{sim.birth.rates.base}},
\code{\link{sim.birth.rates.distn}}, \code{\link{sim.birth.rates.level}},
\code{\link{sim.birth.counts.base}}, \code{\link{sim.birth.counts.distn}},
\code{\link{sim.birth.counts.level}}, and \code{\link{sim.birth.standard}}.
For a version with cell sizes of 30 rather than 300, see
\code{\link{sim.fem.popn.small}}.
}
\keyword{datasets}
