% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fem.popn.small.R
\docType{data}
\name{sim.fem.popn.small}
\alias{sim.fem.popn.small}
\title{Simulated data on women of childbearing age.}
\format{An array with dimensions \code{"age"}, \code{"region"},
and \code{"time"}.}
\usage{
sim.fem.popn.small
}
\description{
Simulated numbers of women of childbearin age, by 5-year age group,
30 regions, and 3 time period.  Every cell in the the array has a
value of 30.
}
\seealso{
The data are designed for use with
\code{\link{sim.birth.rates.base}},
\code{\link{sim.birth.rates.distn}},
\code{\link{sim.birth.rates.level}},
\code{\link{sim.birth.counts.base.small}},
\code{\link{sim.birth.counts.distn.small}},
\code{\link{sim.birth.counts.level.small}}, and
\code{\link{sim.birth.standard}}.
For a version with cell sizes of 300 rather than 30, see
\code{\link{sim.fem.popn}}.
}
\keyword{datasets}
